src/test/java/com/numericalmethod/suanshu/optimization/constrained/integer/IPProblemImpl1Test.java
/*
 * Copyright (c) Numerical Method Inc.
 * http://www.numericalmethod.com/
 * 
 * THIS SOFTWARE IS LICENSED, NOT SOLD.
 * 
 * YOU MAY USE THIS SOFTWARE ONLY AS DESCRIBED IN THE LICENSE.
 * IF YOU ARE NOT AWARE OF AND/OR DO NOT AGREE TO THE TERMS OF THE LICENSE,
 * DO NOT USE THIS SOFTWARE.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITH NO WARRANTY WHATSOEVER,
 * EITHER EXPRESS OR IMPLIED, INCLUDING, WITHOUT LIMITATION,
 * ANY WARRANTIES OF ACCURACY, ACCESSIBILITY, COMPLETENESS,
 * FITNESS FOR A PARTICULAR PURPOSE, MERCHANTABILITY, NON-INFRINGEMENT, 
 * TITLE AND USEFULNESS.
 * 
 * IN NO EVENT AND UNDER NO LEGAL THEORY,
 * WHETHER IN ACTION, CONTRACT, NEGLIGENCE, TORT, OR OTHERWISE,
 * SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR
 * ANY CLAIMS, DAMAGES OR OTHER LIABILITIES,
 * ARISING AS A RESULT OF USING OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.numericalmethod.suanshu.optimization.constrained.integer;

import com.numericalmethod.suanshu.analysis.function.rn2r1.RealScalarFunction;
import com.numericalmethod.suanshu.vector.doubles.Vector;
import org.junit.*;
import static org.junit.Assert.*;

/**
 *
 * @author Haksun Li
 */
public class IPProblemImpl1Test {

    @Test
    public void test_0010() {
        IPProblemImpl1 instance = new IPProblemImpl1(
                new RealScalarFunction() {

                    public Double evaluate(Vector x) {
                        return x.get(1) + x.get(2) + x.get(3);
                    }

                    public int dimensionOfDomain() {
                        return 3;
                    }

                    public int dimensionOfRange() {
                        return 1;
                    }
                },
                null, null,
                new int[]{2, 3},
                0.0001);

        assertArrayEquals(new int[]{2, 3}, instance.getIntegerIndices());
        assertArrayEquals(new int[]{3}, instance.getNonIntegralIndices(new double[]{0.1, 2.00001, 3.01}));
    }
}
